+++ CHANGELOG.md
@@ -78,6 +78,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 * [#756](https://github.com/allora-network/allora-chain/pull/756) Prevent ecosystem rewards after its supply hit
+* [#758](https://github.com/allora-network/allora-chain/pull/758): Ensure pending stake removals are processed
 * [#767](https://github.com/allora-network/allora-chain/pull/767) Fix [ISA-2025-001](https://github.com/cosmos/ibc-go/security/advisories/GHSA-4wf3-5qj9-368v)
 
 # [Released]